+++ egs/mobvoi/v1/run_e2e.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
+# Copyright 2018-2020  Daniel Povey
+#           2018-2020  Yiming Wang
+
+# This recipe uses E2E LF-MMI training which doesn't require GMM training to obtain alignments.
+# Its performance is slightly better than those based on alignments (cross-entropy or regular LF-MMI)
+# on this dataset.
 
 stage=0
 
@@ -43,7 +49,9 @@ if [ $stage -le 4 ]; then
   echo "$0: Post processing transcripts..."
   for folder in train dev eval; do
     dir=data/$folder
+    export LC_ALL=en_US.UTF-8
     cat $dir/text | awk '{if ($2=="嗨小问" || $2=="嗨小问嗨小问") {print $1,"嗨小问";} else {print $1,"FREETEXT"}}' > $dir/text.tmp || exit 1
+    export LC_ALL=C
     cat $dir/text.tmp > $dir/text || exit 1
     rm -f $dir/text.tmp 2>/dev/null || true
   done
@@ -60,7 +68,9 @@ fi
 if [ $stage -le 6 ]; then
   id_sil=`cat data/lang/words.txt | grep "<sil>" | awk '{print $2}'`
   id_freetext=`cat data/lang/words.txt | grep "FREETEXT" | awk '{print $2}'`
+  export LC_ALL=en_US.UTF-8
   id_word=`cat data/lang/words.txt | grep "嗨小问" | awk '{print $2}'`
+  export LC_ALL=C
   mkdir -p data/lang/lm
   cat <<EOF > data/lang/lm/fst.txt
 0 1 $id_sil $id_sil
@@ -142,15 +152,12 @@ if [ $stage -le 8 ]; then
   done
 
   # Augment with musan_noise
-  export LC_ALL=en_US.UTF-8 
-  steps/data/augment_data_dir_for_asr.py --utt-prefix "noise" --fg-interval 1 --fg-snrs "15:10:5:0" --fg-noise-dir "data/musan_noise" data/train_shorter data/train_shorter_noise
-  cat data/train_shorter/utt2dur | awk -v name=noise '{print name"_"$0}' >data/train_shorter_noise/utt2dur
+  export LC_ALL=en_US.UTF-8
+  steps/data/augment_data_dir.py --utt-prefix "noise" --modify-spk-id true --fg-interval 1 --fg-snrs "15:10:5:0" --fg-noise-dir "data/musan_noise" data/train_shorter data/train_shorter_noise
   # Augment with musan_music
-  steps/data/augment_data_dir_for_asr.py --utt-prefix "music" --bg-snrs "15:10:8:5" --num-bg-noises "1" --bg-noise-dir "data/musan_music" data/train_shorter data/train_shorter_music
-  cat data/train_shorter/utt2dur | awk -v name=music '{print name"_"$0}' >data/train_shorter_music/utt2dur
+  steps/data/augment_data_dir.py --utt-prefix "music" --modify-spk-id true --bg-snrs "15:10:8:5" --num-bg-noises "1" --bg-noise-dir "data/musan_music" data/train_shorter data/train_shorter_music
   # Augment with musan_speech
-  steps/data/augment_data_dir_for_asr.py --utt-prefix "babble" --bg-snrs "20:17:15:13" --num-bg-noises "3:4:5:6:7" --bg-noise-dir "data/musan_speech" data/train_shorter data/train_shorter_babble
-  cat data/train_shorter/utt2dur | awk -v name=babble '{print name"_"$0}' >data/train_shorter_babble/utt2dur
+  steps/data/augment_data_dir.py --utt-prefix "babble" --modify-spk-id true --bg-snrs "20:17:15:13" --num-bg-noises "3:4:5:6:7" --bg-noise-dir "data/musan_speech" data/train_shorter data/train_shorter_babble
   export LC_ALL=C
 fi
