+++ egs/snips/v1/conf/online_cmvn.conf
@@ -1,1 +1,3 @@
 # configuration file for apply-cmvn-online, used in the script ../local/run_online_decoding.sh
+--norm-means=true
+--norm-vars=false
