--- egs/snips/v1/local/compute_min_dcf.py
@@ -1,1 +1,0 @@
-../../../mobvoi/v1/local/compute_min_dcf.py
