+++ egs/snips/v1/local/get_random_subsegments.py
@@ -1,1 +1,1 @@
-../../../mobvoi/v1/local/get_random_subsegments.py
+../../../../scripts/wakeword/get_random_subsegments.py
