--- egs/snips/v1/local/parse_cost.py
@@ -1,1 +1,0 @@
-../../../mobvoi/v1/local/parse_cost.py
