--- egs/snips/v1/local/plot_scatter.py
@@ -1,1 +1,0 @@
-../../../mobvoi/v1/local/plot_scatter.py
