--- egs/snips/v1/local/process_lattice.sh
@@ -1,1 +1,0 @@
-../../../mobvoi/v1/local/process_lattice.sh
