--- egs/snips/v1/local/score.sh
@@ -1,1 +1,0 @@
-../../../mobvoi/v1/local/score.sh
