+++ egs/snips/v1/run.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+# Copyright 2018-2020  Daniel Povey
+#           2018-2020  Yiming Wang
 
 stage=0
 
@@ -133,14 +135,11 @@ if [ $stage -le 6 ]; then
   done
 
   # Augment with musan_noise
-  steps/data/augment_data_dir_for_asr.py --utt-prefix "noise" --fg-interval 1 --fg-snrs "15:10:5:0" --fg-noise-dir "data/musan_noise" data/train_shorter data/train_shorter_noise
-  cat data/train_shorter/utt2dur | awk -v name=noise '{print name"_"$0}' >data/train_shorter_noise/utt2dur
+  steps/data/augment_data_dir.py --utt-prefix "noise" --fg-interval 1 --fg-snrs "15:10:5:0" --fg-noise-dir "data/musan_noise" data/train_shorter data/train_shorter_noise
   # Augment with musan_music
-  steps/data/augment_data_dir_for_asr.py --utt-prefix "music" --bg-snrs "15:10:8:5" --num-bg-noises "1" --bg-noise-dir "data/musan_music" data/train_shorter data/train_shorter_music
-  cat data/train_shorter/utt2dur | awk -v name=music '{print name"_"$0}' >data/train_shorter_music/utt2dur
+  steps/data/augment_data_dir.py --utt-prefix "music" --bg-snrs "15:10:8:5" --num-bg-noises "1" --bg-noise-dir "data/musan_music" data/train_shorter data/train_shorter_music
   # Augment with musan_speech
-  steps/data/augment_data_dir_for_asr.py --utt-prefix "babble" --bg-snrs "20:17:15:13" --num-bg-noises "3:4:5:6:7" --bg-noise-dir "data/musan_speech" data/train_shorter data/train_shorter_babble
-  cat data/train_shorter/utt2dur | awk -v name=babble '{print name"_"$0}' >data/train_shorter_babble/utt2dur
+  steps/data/augment_data_dir.py --utt-prefix "babble" --bg-snrs "20:17:15:13" --num-bg-noises "3:4:5:6:7" --bg-noise-dir "data/musan_speech" data/train_shorter data/train_shorter_babble
 fi
 
 if [ $stage -le 7 ]; then
@@ -171,7 +170,6 @@ if [ $stage -le 8 ]; then
   steps/align_si.sh --nj 20 --cmd "$train_cmd" \
     data/train_shorter data/lang exp/mono exp/mono_ali_train_shorter
 fi
-
 
 if [ $stage -le 9 ]; then
   echo "$0: preparing for low-resolution speed-perturbed data (for alignment)"
